# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Madrissti educational prototype",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/demo/auth": {
      "post": {
        "summary": "Auth",
        "operationId": "auth_v1_demo_auth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/courses": {
      "get": {
        "summary": "Courses",
        "operationId": "courses_v1_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/courses/{course_id}/lessons": {
      "get": {
        "summary": "Catalogue",
        "operationId": "catalogue_v1_courses__course_id__lessons_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lessons/{lesson_id}": {
      "get": {
        "summary": "Lesson",
        "operationId": "lesson_v1_lessons__lesson_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "summary": "Start",
        "operationId": "start_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Selection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/sessions/{sid}/context": {
      "post": {
        "summary": "Context",
        "operationId": "context_v1_sessions__sid__context_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Context"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{sid}/workspace/fraction-bars": {
      "post": {
        "summary": "Render",
        "operationId": "render_v1_sessions__sid__workspace_fraction_bars_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{sid}/realtime": {
      "post": {
        "summary": "Realtime",
        "operationId": "realtime_v1_sessions__sid__realtime_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{sid}/summary": {
      "get": {
        "summary": "Summary",
        "operationId": "summary_v1_sessions__sid__summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{sid}/end": {
      "post": {
        "summary": "End",
        "operationId": "end_v1_sessions__sid__end_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{sid}": {
      "delete": {
        "summary": "Delete",
        "operationId": "delete_v1_sessions__sid__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Auth": {
        "properties": {
          "access_code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Access Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_code"
        ],
        "title": "Auth"
      },
      "Context": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "title": "Lesson Id"
          },
          "content_version": {
            "type": "string",
            "title": "Content Version"
          },
          "selected_block_id": {
            "type": "string",
            "title": "Selected Block Id"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ar"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "text"
          },
          "preferences": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Preferences"
          },
          "previous_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Previous Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lesson_id",
          "content_version",
          "selected_block_id",
          "previous_revision"
        ],
        "title": "Context"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Rational": {
        "properties": {
          "numerator": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Numerator"
          },
          "denominator": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Denominator"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "numerator",
          "denominator"
        ],
        "title": "Rational"
      },
      "RenderRequest": {
        "properties": {
          "whole_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Whole Id"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/Rational"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Values"
          },
          "context_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Context Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "whole_id",
          "values",
          "context_revision"
        ],
        "title": "RenderRequest"
      },
      "Selection": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "title": "Lesson Id"
          },
          "content_version": {
            "type": "string",
            "title": "Content Version"
          },
          "selected_block_id": {
            "type": "string",
            "title": "Selected Block Id"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ar"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "text"
          },
          "preferences": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Preferences"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lesson_id",
          "content_version",
          "selected_block_id"
        ],
        "title": "Selection"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}